refactor: extract home screen components (closes #39)
Split apps/mobile/app/(app)/index.tsx (749L → 221L) into focused
units:

- hooks/useActiveRound — useFocusEffect-based query for the resume
  banner; encapsulates the rounds + hole_scores fetch + max-hole
  derivation
- components/home/ResumeRoundBanner — pulsing border + Resume link;
  pulse worklet self-cleans on unmount
- components/home/SGBreakdown — diverging SG bars with their own
  breakdown calc + maxAbs + SGBar primitive
- components/home/SGTrendChart — VictoryChart wrap with axis styling
- components/home/RecentRoundsList — swipe-to-delete list, empty
  state, see-all link, SGValue primitive
- components/home/RoundCTAs — Start live + Log past buttons

index.tsx stays the orchestrator: profile fetch, sync queue, recent
SG fetch, delete confirmation. Pure extraction — no behavior change.
